added check of ssh-key
